requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-openai
mcp-use
pandas
elasticsearch>=8.0.0
kaggle
httpx

